requirements.txt
Flask==0.12.2
psycopg2-binary